refactor: consolidate shared utilities, centralize CLI flags, reorganize into subdirectories
- Extract AES-128 implementation to standalone header (Data/AES128.h)
- Move shared helpers to Utils.h/cpp: collectStringGlobals, getOrDeclare,
  buildByteArrayConstant, createPrivateByteGlobal, fillRandomBytes
- Create Options.h/cpp: single source of truth for all -kagura-* CLI flags,
  eliminating scattered cl::opt definitions and extern chains across 12 files
- Reorganize lib/Transforms/ into logical subdirectories:
  CFG/, Data/, AntiAnalysis/, Platform/
- Simplify StringEncryptionAES.cpp from 447 to 180 lines using shared utils
- Remove duplicate code from StringEncryption.cpp, AntiDebug.cpp, AntiTamper.cpp
- Improve README.md with architecture diagram, badges, and clearer structure
