+++ lib/Transforms/Data/AES128.h
@@ -1,0 +1,138 @@
+#pragma once
+//===-- AES128.h - Compile-time AES-128 implementation --------------------===//
+//
+// Table-based Rijndael forward cipher and CTR mode, used by
+// StringEncryptionAES to encrypt strings at compile time.
+//
+//===----------------------------------------------------------------------===//
+
+#include <array>
+#include <cstdint>
+#include <cstring>
+#include <vector>
+
+namespace kagura {
+namespace aes {
+
+// Rijndael S-box
+static constexpr uint8_t SBox[256] = {
+  0x63,0x7c,0x77,0x7b,0xf2,0x6b,0x6f,0xc5,0x30,0x01,0x67,0x2b,0xfe,0xd7,0xab,0x76,
+  0xca,0x82,0xc9,0x7d,0xfa,0x59,0x47,0xf0,0xad,0xd4,0xa2,0xaf,0x9c,0xa4,0x72,0xc0,
+  0xb7,0xfd,0x93,0x26,0x36,0x3f,0xf7,0xcc,0x34,0xa5,0xe5,0xf1,0x71,0xd8,0x31,0x15,
+  0x04,0xc7,0x23,0xc3,0x18,0x96,0x05,0x9a,0x07,0x12,0x80,0xe2,0xeb,0x27,0xb2,0x75,
+  0x09,0x83,0x2c,0x1a,0x1b,0x6e,0x5a,0xa0,0x52,0x3b,0xd6,0xb3,0x29,0xe3,0x2f,0x84,
+  0x53,0xd1,0x00,0xed,0x20,0xfc,0xb1,0x5b,0x6a,0xcb,0xbe,0x39,0x4a,0x4c,0x58,0xcf,
+  0xd0,0xef,0xaa,0xfb,0x43,0x4d,0x33,0x85,0x45,0xf9,0x02,0x7f,0x50,0x3c,0x9f,0xa8,
+  0x51,0xa3,0x40,0x8f,0x92,0x9d,0x38,0xf5,0xbc,0xb6,0xda,0x21,0x10,0xff,0xf3,0xd2,
+  0xcd,0x0c,0x13,0xec,0x5f,0x97,0x44,0x17,0xc4,0xa7,0x7e,0x3d,0x64,0x5d,0x19,0x73,
+  0x60,0x81,0x4f,0xdc,0x22,0x2a,0x90,0x88,0x46,0xee,0xb8,0x14,0xde,0x5e,0x0b,0xdb,
+  0xe0,0x32,0x3a,0x0a,0x49,0x06,0x24,0x5c,0xc2,0xd3,0xac,0x62,0x91,0x95,0xe4,0x79,
+  0xe7,0xc8,0x37,0x6d,0x8d,0xd5,0x4e,0xa9,0x6c,0x56,0xf4,0xea,0x65,0x7a,0xae,0x08,
+  0xba,0x78,0x25,0x2e,0x1c,0xa6,0xb4,0xc6,0xe8,0xdd,0x74,0x1f,0x4b,0xbd,0x8b,0x8a,
+  0x70,0x3e,0xb5,0x66,0x48,0x03,0xf6,0x0e,0x61,0x35,0x57,0xb9,0x86,0xc1,0x1d,0x9e,
+  0xe1,0xf8,0x98,0x11,0x69,0xd9,0x8e,0x94,0x9b,0x1e,0x87,0xe9,0xce,0x55,0x28,0xdf,
+  0x8c,0xa1,0x89,0x0d,0xbf,0xe6,0x42,0x68,0x41,0x99,0x2d,0x0f,0xb0,0x54,0xbb,0x16,
+};
+
+static constexpr uint8_t Rcon[11] = {
+  0x00,0x01,0x02,0x04,0x08,0x10,0x20,0x40,0x80,0x1b,0x36,
+};
+
+inline uint8_t gmul(uint8_t a, uint8_t b) {
+  uint8_t p = 0;
+  for (int i = 0; i < 8; ++i) {
+    if (b & 1) p ^= a;
+    bool hi = (a & 0x80) != 0;
+    a = static_cast<uint8_t>(a << 1);
+    if (hi) a ^= 0x1b;
+    b >>= 1;
+  }
+  return p;
+}
+
+using Block = std::array<uint8_t, 16>;
+using RoundKeys = std::array<Block, 11>;
+
+inline RoundKeys expandKey(const uint8_t key[16]) {
+  RoundKeys rk;
+  for (int i = 0; i < 16; ++i)
+    rk[0][i] = key[i];
+
+  for (int r = 1; r <= 10; ++r) {
+    uint8_t t0 = SBox[rk[r-1][13]] ^ Rcon[r];
+    uint8_t t1 = SBox[rk[r-1][14]];
+    uint8_t t2 = SBox[rk[r-1][15]];
+    uint8_t t3 = SBox[rk[r-1][12]];
+
+    rk[r][0]  = rk[r-1][0]  ^ t0;
+    rk[r][1]  = rk[r-1][1]  ^ t1;
+    rk[r][2]  = rk[r-1][2]  ^ t2;
+    rk[r][3]  = rk[r-1][3]  ^ t3;
+    for (int c = 1; c < 4; ++c) {
+      rk[r][c*4+0] = rk[r-1][c*4+0] ^ rk[r][c*4-4];
+      rk[r][c*4+1] = rk[r-1][c*4+1] ^ rk[r][c*4-3];
+      rk[r][c*4+2] = rk[r-1][c*4+2] ^ rk[r][c*4-2];
+      rk[r][c*4+3] = rk[r-1][c*4+3] ^ rk[r][c*4-1];
+    }
+  }
+  return rk;
+}
+
+inline Block encryptBlock(Block in, const RoundKeys &rk) {
+  Block s = in;
+  for (int i = 0; i < 16; ++i) s[i] ^= rk[0][i];
+
+  for (int r = 1; r <= 9; ++r) {
+    for (int i = 0; i < 16; ++i) s[i] = SBox[s[i]];
+    // ShiftRows
+    uint8_t tmp = s[1];
+    s[1] = s[5]; s[5] = s[9]; s[9] = s[13]; s[13] = tmp;
+    std::swap(s[2], s[10]); std::swap(s[6], s[14]);
+    tmp = s[15]; s[15] = s[11]; s[11] = s[7]; s[7] = s[3]; s[3] = tmp;
+    // MixColumns
+    for (int c = 0; c < 4; ++c) {
+      uint8_t a0 = s[c*4+0], a1 = s[c*4+1], a2 = s[c*4+2], a3 = s[c*4+3];
+      s[c*4+0] = gmul(0x02,a0)^gmul(0x03,a1)^a2^a3;
+      s[c*4+1] = a0^gmul(0x02,a1)^gmul(0x03,a2)^a3;
+      s[c*4+2] = a0^a1^gmul(0x02,a2)^gmul(0x03,a3);
+      s[c*4+3] = gmul(0x03,a0)^a1^a2^gmul(0x02,a3);
+    }
+    for (int i = 0; i < 16; ++i) s[i] ^= rk[r][i];
+  }
+
+  // Round 10
+  for (int i = 0; i < 16; ++i) s[i] = SBox[s[i]];
+  uint8_t tmp = s[1];
+  s[1] = s[5]; s[5] = s[9]; s[9] = s[13]; s[13] = tmp;
+  std::swap(s[2], s[10]); std::swap(s[6], s[14]);
+  tmp = s[15]; s[15] = s[11]; s[11] = s[7]; s[7] = s[3]; s[3] = tmp;
+  for (int i = 0; i < 16; ++i) s[i] ^= rk[10][i];
+
+  return s;
+}
+
+/// AES-128 CTR encrypt/decrypt (symmetric).
+inline std::vector<uint8_t> ctrCrypt(const uint8_t *data, size_t len,
+                                     const uint8_t key[16],
+                                     const uint8_t nonce[8]) {
+  RoundKeys rk = expandKey(key);
+  std::vector<uint8_t> out(len);
+
+  for (size_t offset = 0; offset < len; offset += 16) {
+    uint64_t counter = static_cast<uint64_t>(offset / 16);
+    Block ctrBlock;
+    for (int i = 0; i < 8; ++i) ctrBlock[i] = nonce[i];
+    for (int i = 0; i < 8; ++i)
+      ctrBlock[8 + i] = static_cast<uint8_t>((counter >> ((7 - i) * 8)) & 0xFF);
+
+    Block ks = encryptBlock(ctrBlock, rk);
+
+    size_t blockLen = std::min(static_cast<size_t>(16), len - offset);
+    for (size_t j = 0; j < blockLen; ++j)
+      out[offset + j] = data[offset + j] ^ ks[j];
+  }
+  return out;
+}
+
+} // namespace aes
+} // namespace kagura
