+++ lib/Transforms/Platform/JNIObfuscation.cpp
