+++ lib/Transforms/Platform/VMObfuscation.cpp
